Add user creation

+++ src/android_app/app/src/main/java/com/mobilehealthsports/vaccinepass/business/database/persistence/DbUser.kt
@@ -7,7 +7,7 @@ import java.util.*
 
 @Entity(tableName = "user")
 data class DbUser(
-        @PrimaryKey(autoGenerate = true) val uid: Int,
+        @PrimaryKey(autoGenerate = true) val uid: Long,
         @ColumnInfo(name = "first_name") val firstName: String?,
         @ColumnInfo(name = "last_name") val lastName: String?,
         @ColumnInfo(name = "blood_type") val bloodType: String?,
