+++ src/android_app/app/src/main/java/com/mobilehealthsports/vaccinepass/business/models/User.kt
@@ -6,14 +6,14 @@ import java.io.Serializable
 import java.util.*
 
 data class User(
-        val uid: Int,
-        val firstName: String?,
-        val lastName: String?,
-        val bloodType: String?,
-        val birthDay: Date?,
-        val weight: Float?,
-        val height: Float?,
-        val themeColor: Int,
+    val uid: Long,
+    val firstName: String?,
+    val lastName: String?,
+    val bloodType: String?,
+    val birthDay: Date?,
+    val weight: Float?,
+    val height: Float?,
+    val themeColor: Int,
 ) : Serializable {
     fun resolveColor(context: Context): Int {
         return ContextCompat.getColor(context, themeColor)
